Audio + mods to PlayerWeapon
Instead of the enemy detecting an object and taking damage, I've added a
way for the PlayerWeapon script to damage the enemy. I also created a
parent class, "Enemy", from which I plan to derive all the bad guys. I
also added a great script for audio created by Brackeys.