Rhino optimization with BigRaycastController.cs
I realized that the rhino had the same space between raycasts as the
player to detect collisions. Because the rhino is so big, this was using
a lot of resources, so I created a different script for larger raycast
controllers. This cut the rhino's resource hoggin by more than a half.