Individual ice legs do raycasting and fix the PointAtPlayer IEnumerator
So now the individual legs do the raycasting. This means that I can set individual raylengths for each leg. It's better. Also, the PointAtPlayer thing was holding up. I'm not sure why exactly, but I'm now using euler angles in the condition in the while loop and adding a small buffer on either side of the angle, which offers a little more flexibility.